添加移动采集分析结果

+++ src/main/java/com/sensenets/sinopec/buiness/service/ICollectResultFlowService.java
@@ -26,7 +26,7 @@ public interface ICollectResultFlowService {
 
     int updateByExample(CollectResultFlow record, CollectResultFlowCriteria example);
 
-    int insert(CollectResultFlow record);
+    long insert(CollectResultFlow record);
 
     int insertSelective(CollectResultFlow record);
 }
